refactory: store 수정
