+++ src/store/index.js
@@ -1,1 +1,2 @@
 export { store } from 'store/store';
+export { setAllUsers, setUsersInfo, setPostsInfo } from 'store/searchSlice'
